clean up and separate matchers
